refactor: 화면 너비에 따른 반응형타입 hook 및 ProgressBar 코드에 적용
